requirements.txt
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
pytest>=7.0.0
pytest-cov>=4.0.0 